bgpd: Don't let path ordering impact if a path is considered multipath.
In each of the following scenarios, we won't consider a path as multipath:
 your secondary bestpath is first
 your bestpath is last
 the bestpath and secondary bestpath are not similar enough for multipath
 a bunch of multipath eligible paths are in the middle

When you hit this scenario we don't ECMP over any of those paths in the middle
of the list.

We need to figure out what our bestpath is first and then come back and
find/tag all of the multipaths. Thus, we use two passes to determine if
a path is multipath or not.

Signed-off-by: Daniel Walton <dwalton@cumulusnetworks.com>
Reviewed-by: Dinesh Dutt <ddutt@cumulusnetworks.com>