bgpd: Reset keepalive timer when an actual update or keepalive is sent
When the keepalive timer expires, the existing code checks if there are
any updates to be sent (bgp_routeq_empty), and if there are, it decides
to not send a keepalive. When fast keepalive and hold timers are used,
this can be a problem, since updates are themselves sent based on another
timer, the advertisement interval timer, among other things. Therefore,
the hold timer can expire on a peer causing session resets. To avoid this,
we reset the keepalive timer when a packet is actually transmitted to a
peer. This enables efficient keepalive sends, while accounting for batching
efficiencies in sending updates.

Signed-off-by: Dinesh Dutt <ddutt@cumulusnetworks.com>
Signef-off-by: Daniel Walton <dwalton@cumulusnetworks.com>