bgpd-handle-incoming-connection-for-inactive-peer.patch
When an incoming connection is received from a neighbor that is configured but
is not activated for any address-family, the connection is accepted without
taking further action. This causes the connection to hang in OpenSent on the
neighbor and can in turn delay the connection setup. Fix to reject incoming
connections when there is no address-family activated for the neighbor.

Signed-off-by: Vivek Venkataraman <vivek@cumulusnetworks.com>