When a peering is being established, the IPv4 and IPv6 addresses of the local end of the connection, as applicable, are obtained and stored in the peer's 'nexthop' structure to facilitate filling of the NEXT_HOP field in Update messages among other things. The process of obtaining the local address involves examination of the list of interfaces to identify a match corresponding to the socket address of the connection.
There are timing conditions, especially when BGPD starts with a config,
where the interface may not have reached BGP from Zebra at the time a
peering reaches the state to determine the local addresses. The code does
not handle this well and the result could be Updates generated with bad
(Martian) NEXT_HOP values. Resolve the issue by bringing down the connection
in this case as not identifying the local addresses is really an error.