BGP: Handle connected address notifications correctly when using Unnumbered
When we're trying to use BGP unnumbered and the interface has a non-/30 or
/31 address, when we receive notification of the interface having an IPv4
address, we attempt to reinitialize the BGP FSM which causes the peer to
never come up if its attempting to use IPv6 link-local peering.

To fix this, when we restart the BGP FSM, we verify if the peer has a valid
IP address (v4 or v6). If it does not, the FSM does not proceed but stays in
the same startup state allowing the correct startup when the peer's LL address
is learnt.

Signed-off-by: Dinesh G Dutt <ddutt@cumulusnetworks.com>