Zebra: Fix replace route for uninstall scenario
When a Quagga route that is currently installed is superceded by a
kernel route (connected or static route to same destination), the
Quagga route is not uninstalled from the kernel. This would be an
issue only in this kind of scenario, one example of which is
Anycast IP used for VxLAN A-A.

Fix by ensuring this case is handled correctly.