Vtysh: Fix configuration grouping in case of integrated config file
When creating an integrated configuration file, certain IPv6 commands
would end up bunched with the wrong set of commands instead of being
with its IPv4 compatriot. An example of this was the newly added
ipv6 protocol filter command which ended up at the head of the config
file mixed in with log commands instead of its IPv4 equivalent cmd.
When the configuration was saved in individual files, this issue was
not seen.

This patch addresses this inconsistency by fixing the grouping
of commands for integrated config files. For integrated config files,
there exists a function that attempts to determine which group of
commands belong together. That function had to be updated for some
of the newly added commands.

Signed-off-by: Dinesh G Dutt <ddutt@cumulusnetworks.com>