The IPv6 OSPF BFD sessions stay down after ifdown/ifup on logical interfaces.
When the interface is brought down the IPv6 neighbors discovered on that interface are deleted. This deletion happens without first bringing down the neighbor and the BFD deregistration happens only when the neighbor state changes. This leaves an orphaned BFD session in PTM. Also, the BFD session socket that is bound to the interface that was brought down loses connection. The socket has to be rebound to the interface when it comes up. This problem will not happen if the client deleted the sessions and re-adds it when interface goes down and comes up.

Fix: Added the BFD deregistration code to IPv6 OSPF neighbor delete.

    Signed-off-by: Radhika Mahankali <radhika@cumulusnetworks.com>
    Reviewed-by:   Donald Sharp <sharpd@cumulusnetworks.com>